Removing the 'Vnext' suffix from MSFButtonVnext and associated types.
